changed file names for clarity SpendingByRecipient --> SpendingByRecipientContainer SpendingByRecipientContainer --> RecipientTableContainer
