+++ src/js/containers/covid19/recipient/SpendingByRecipientContainer.jsx
@@ -3,345 +3,137 @@
  * Created by Lizzie Salita 7/8/20
  */
 
-import React, { useState, useEffect, useCallback, useRef } from 'react';
-import PropTypes from 'prop-types';
+import React, { useState, useEffect, useRef } from 'react';
 import { useSelector } from 'react-redux';
-import { isCancel } from 'axios';
-import replaceString from 'helpers/replaceString';
-import { Table, Pagination, SearchBar } from 'data-transparency-ui';
-import { Link } from 'react-router-dom';
 
+import { awardTypeTabs } from 'dataMapping/covid19/covid19';
 import { awardTypeGroups } from 'dataMapping/search/awardType';
-import BaseSpendingByRecipientRow from 'models/v2/covid19/BaseSpendingByRecipientRow';
-import { spendingTableSortFields } from 'dataMapping/covid19/covid19';
-import { fetchDisasterSpending, fetchLoanSpending } from 'apis/disaster';
-import Note from 'components/sharedComponents/Note';
-import noteText from 'dataMapping/covid19/recipient/recipient';
-import TableDownloadLink from 'containers/covid19/TableDownloadLink';
+import { fetchDisasterSpendingCount } from 'apis/disaster';
+import { areCountsDefined } from 'helpers/covid19Helper';
+import SummaryInsightsContainer from 'containers/covid19/SummaryInsightsContainer';
+import RecipientTableContainer from 'containers/covid19/recipient/RecipientTableContainer';
+import GlossaryLink from 'components/sharedComponents/GlossaryLink';
 import Analytics from 'helpers/analytics/Analytics';
-import { calculateUnlinkedTotals } from 'helpers/covid19Helper';
+import { Tabs } from "data-transparency-ui";
+import { scrollIntoView } from 'containers/covid19/helpers/scrollHelper';
 import { useStateWithPrevious } from 'helpers';
 
-const propTypes = {
-    activeTab: PropTypes.string.isRequired,
-    prevActiveTab: PropTypes.string,
-    scrollIntoView: PropTypes.func.isRequired
-};
-
-const columns = [
+const overviewData = [
     {
-        title: 'name',
-        displayName: 'Recipient'
+        type: 'resultsCount',
+        title: 'Number of Recipients'
     },
     {
-        title: 'obligation',
-        displayName: 'Award Obligations',
-        right: true
-    },
-    {
-        title: 'outlay',
-        displayName: 'Award Outlays',
-        right: true
-    },
-    {
-        title: 'awardCount',
-        displayName: (
-            <div className="table-header-label__title">
-                <div>Number</div>
-                <div>of Awards</div>
+        type: 'awardObligations',
+        title: (
+            <div>
+                <span className="glossary-term">Award Obligations</span> <GlossaryLink term="obligation" />
             </div>
         ),
-        right: true
+        isMonetary: true
+    },
+    {
+        type: 'awardOutlays',
+        title: (
+            <div>
+                <span className="glossary-term">Award Outlays</span> <GlossaryLink term="outlay" />
+            </div>
+        ),
+        isMonetary: true
+    },
+    {
+        type: 'numberOfAwards',
+        title: 'Number of Awards'
     }
 ];
 
-const loanColumns = [
-    {
-        title: 'name',
-        displayName: 'Recipient'
-    },
-    {
-        title: 'obligation',
-        displayName: (
-            <div className="table-header-label__title">
-                <div>Award Obligations</div>
-                <div>(Loan Subsidy Cost)</div>
-            </div>
-        ),
-        right: true
-    },
-    {
-        title: 'outlay',
-        displayName: (
-            <div className="table-header-label__title">
-                <div>Award Outlays</div>
-                <div>(Loan Subsidy Cost)</div>
-            </div>
-        ),
-        right: true
-    },
-    {
-        title: 'faceValueOfLoan',
-        displayName: (
-            <div className="table-header-label__title">
-                <div>Face Value</div>
-                <div>of Loans</div>
-            </div>
-        ),
-        right: true
-    },
-    {
-        title: 'awardCount',
-        displayName: (
-            <div className="table-header-label__title">
-                <div>Number</div>
-                <div>of Awards</div>
-            </div>
-        ),
-        right: true
-    }
-];
+const SpendingByRecipientContainer = () => {
+    const [inFlight, setInFlight] = useState(true);
+    const [prevActiveTab, activeTab, setActiveTab] = useStateWithPrevious(awardTypeTabs[0].internal);
+    const { defcParams } = useSelector((state) => state.covid19);
+    const awardFilterButtonsRef = useRef(null);
 
-const clickedRecipientProfile = (recipientName) => {
-    Analytics.event({
-        category: 'COVID-19 - Award Spending by Recipient - Recipients',
-        action: 'recipient profile click',
-        label: recipientName
+    const [tabCounts, setTabCounts] = useState({
+        all: null,
+        grants: null,
+        direct_payments: null,
+        loans: null,
+        other: null,
+        contracts: null,
+        idvs: null
     });
-};
 
-export const parseRows = (rows, activeTab, query) => (
-    rows.map((row) => {
-        const rowData = Object.create(BaseSpendingByRecipientRow);
-        rowData.populate(row);
-        let description = rowData.description;
-        let link = description;
-        if (query) description = replaceString(description, query, "query-matched");
-        if (rowData._childId && rowData._recipientId) {
-            // there are two profile pages for this recipient
-            const handleClick = () => clickedRecipientProfile(`${description}`);
-            link = (
-                <>
-                    {description}&nbsp;(
-                    <Link onClick={handleClick} to={`/recipient/${rowData._childId}/latest`}>
-                        as Child
-                    </Link>,&nbsp;
-                    <Link onClick={handleClick} to={`/recipient/${rowData._recipientId}/latest`}>
-                        as Recipient
-                    </Link>
-                    )
-                </>
-            );
-        }
-        else if (rowData._childId || rowData._recipientId) {
-            const handleClick = clickedRecipientProfile(`${description}`);
-            // there is a single profile page for this recipient
-            link = (
-                <Link onClick={handleClick} to={`/recipient/${rowData._childId || rowData._recipientId}/latest`}>
-                    {description}
-                </Link>
-            );
-        }
-        if (activeTab === 'loans') {
-            return [
-                link,
-                rowData.obligation,
-                rowData.outlay,
-                rowData.faceValueOfLoan,
-                rowData.awardCount
-            ];
-        }
-        return [
-            link,
-            rowData.obligation,
-            rowData.outlay,
-            rowData.awardCount
-        ];
-    })
-);
-
-const SpendingByRecipientContainer = ({ activeTab, prevActiveTab, scrollIntoView }) => {
-    const [currentPage, changeCurrentPage] = useState(1);
-    const [prevPageSize, pageSize, changePageSize] = useStateWithPrevious(10);
-    const [totalItems, setTotalItems] = useState(0);
-    const [results, setResults] = useState([]);
-    const [parsedRows, setParsedRows] = useState([]); // finalized rows (including Unlinked Data)
-    const [resultTotal, setResultTotal] = useState(0);
-    const [loading, setLoading] = useState(true);
-    const [error, setError] = useState(false);
-    const [prevSort, sort, setSort] = useStateWithPrevious('obligation');
-    const [prevOrder, order, setOrder] = useStateWithPrevious('desc');
-    const [prevQuery, query, setQuery] = useStateWithPrevious('');
-    const tableRef = useRef(null);
-    const tableWrapperRef = useRef(null);
-    const errorOrLoadingWrapperRef = useRef(null);
-    const request = useRef(null);
-    const [unlinkedDataClass, setUnlinkedDataClass] = useState(false);
-    const { recipientTotals, defcParams } = useSelector((state) => state.covid19);
-
-    const updateSort = (field, direction) => {
-        setSort(field);
-        setOrder(direction);
+    const changeActiveTab = (tab) => {
+        const selectedTab = awardTypeTabs.find((item) => item.internal === tab).internal;
+        setActiveTab(selectedTab);
+        Analytics.event({ category: 'COVID-19 - Award Spending by Recipient - Recipients', action: `${activeTab} - click` });
     };
 
-    const addUnlinkedData = (rows = results, totals = resultTotal, totalRecipient = recipientTotals) => {
-        // add unlinked data if activeTab is all
-        if (activeTab === 'all' && !query && Object.keys(totalRecipient).length > 0) {
-            const unlinkedData = calculateUnlinkedTotals(totalRecipient, totals);
-            setUnlinkedDataClass(true);
-            const unlinkedName = (
-                <div className="unlinked-data">
-                    Unknown Recipient (Unlinked Data)
-                </div>
-            );
-            const rowData = Object.create(BaseSpendingByRecipientRow);
-
-            rowData.populate({
-                obligation: unlinkedData.obligation,
-                outlay: unlinkedData.outlay,
-                award_count: unlinkedData.award_count
-            });
-
-            const rowsWithUnlinked = rows
-                .concat([[
-                    unlinkedName,
-                    rowData.obligation,
-                    rowData.outlay,
-                    rowData.awardCount
-                ]]);
-            setParsedRows(rowsWithUnlinked);
-        }
-        else {
-            setParsedRows(rows);
-        }
-    };
-
-    const fetchSpendingByRecipientCallback = useCallback(() => {
-        if (request.current) {
-            request.current.cancel();
-        }
-        setLoading(true);
-        if (defcParams && defcParams.length > 0) {
+    useEffect(() => {
+        // Make an API request for the count of Recipients for each award type
+        // Post-MVP this should be updated to use a new endpoint that returns all the counts
+        const promises = awardTypeTabs.map((awardType) => {
             const params = {
                 filter: {
                     def_codes: defcParams
-                },
-                pagination: {
-                    limit: pageSize,
-                    page: currentPage,
-                    sort: spendingTableSortFields[sort],
-                    order
                 }
             };
-            if (activeTab !== 'all') {
-                params.filter.award_type_codes = awardTypeGroups[activeTab];
+            if (awardType.internal !== 'all') {
+                // Endpoint defaults to all award types if award_type_codes is not provided
+                params.filter.award_type_codes = awardTypeGroups[awardType.internal];
             }
-            if (query) {
-                params.filter.query = query;
-            }
-            const recipientRequest = activeTab === 'loans' ? fetchLoanSpending('recipient', params) : fetchDisasterSpending('recipient', params);
-            request.current = recipientRequest;
-            recipientRequest.promise
-                .then((res) => {
-                    const rows = parseRows(res.data.results, activeTab, query);
-                    const totals = res.data.totals;
-                    setResultTotal(totals);
-                    setResults(rows);
-                    setTotalItems(res.data.page_metadata.total);
-                    setLoading(false);
-                    setError(false);
-                }).catch((err) => {
-                    if (!isCancel(err)) {
-                        setError(true);
-                        setLoading(false);
-                        request.current = null;
-                        console.error(err);
-                    }
+            return fetchDisasterSpendingCount('recipient', params).promise;
+        });
+        // Wait for all the requests to complete and then store the results in state
+        Promise.all(promises)
+            .then(([allRes, grantsRes, loansRes, directPaymentsRes, otherRes, contractRes, idvRes]) => {
+                setTabCounts({
+                    all: allRes.data.count,
+                    grants: grantsRes.data.count,
+                    direct_payments: directPaymentsRes.data.count,
+                    loans: loansRes.data.count,
+                    other: otherRes.data.count,
+                    contracts: contractRes.data.count,
+                    idvs: idvRes.data.count
                 });
-        }
-    });
+            });
+    }, [defcParams]);
 
+    const scrollIntoViewTable = (loading, error, errorOrLoadingRef, tableWrapperRef, margin, scrollOptions) => {
+        scrollIntoView(loading, error, errorOrLoadingRef, tableWrapperRef, margin, scrollOptions, awardFilterButtonsRef);
+    };
     useEffect(() => {
-        if (Object.keys(recipientTotals).length && results.length) {
-            addUnlinkedData();
+        const countState = areCountsDefined(tabCounts);
+        if (!countState) {
+            setInFlight(true);
         }
-    }, [recipientTotals, resultTotal, results]);
-
-    useEffect(() => {
-        // Reset to the first page
-        if (currentPage !== 1) {
-            changeCurrentPage(1);
+        else if (countState) {
+            setInFlight(false);
         }
-        else {
-            const hasParamChanged = (
-                prevOrder !== order ||
-                prevSort !== sort ||
-                prevPageSize !== pageSize ||
-                prevQuery !== query ||
-                prevActiveTab !== activeTab
-            );
-            if (hasParamChanged) {
-                fetchSpendingByRecipientCallback();
-            }
-        }
-    }, [pageSize, defcParams, sort, order, activeTab, query]);
-
-    useEffect(() => {
-        fetchSpendingByRecipientCallback();
-    }, [currentPage]);
-
-    useEffect(() => {
-        scrollIntoView(loading, error, errorOrLoadingWrapperRef, tableWrapperRef, 130, true);
-    }, [loading, error]);
+    }, [tabCounts, setInFlight]);
 
     return (
-        <div ref={tableWrapperRef}>
-            <div className="table-utility">
-                <div className="table-utility__left">
-                    <SearchBar onSearch={setQuery} />
-                </div>
-                {(!error && !loading && results.length > 0) &&
-                <div className="table-utility__right">
-                    <TableDownloadLink
-                        defCodes={defcParams && defcParams.length > 0 && defcParams}
-                        awardTypeCodes={awardTypeGroups[activeTab] ? awardTypeGroups[activeTab] : null}
-                        query={query} />
-                </div>}
+        <div className="spending-by-recipient">
+            <div ref={awardFilterButtonsRef}>
+                <Tabs
+                    active={activeTab}
+                    types={awardTypeTabs.map((tab) => ({ ...tab, disabled: tabCounts && tab.internal !== 'all' && !tabCounts[tab.internal], count: tabCounts[tab.internal] }))}
+                    switchTab={changeActiveTab}
+                    tablessStyle />
             </div>
-            {(results.length > 0 || error) && <Pagination
-                currentPage={currentPage}
-                changePage={changeCurrentPage}
-                changeLimit={changePageSize}
-                limitSelector
-                resultsText
-                pageSize={pageSize}
-                totalItems={totalItems} />}
-            <div ref={tableRef} className={`table-wrapper ${unlinkedDataClass ? 'unlinked-data' : ''}`} >
-                <Table
-                    columns={activeTab === 'loans' ? loanColumns : columns}
-                    rows={parsedRows}
-                    updateSort={updateSort}
-                    currentSort={{ field: sort, direction: order }}
-                    error={error}
-                    loading={loading} />
-            </div>
-            {(results.length > 0 || error) && <Pagination
-                currentPage={currentPage}
-                changePage={changeCurrentPage}
-                changeLimit={changePageSize}
-                limitSelector
-                resultsText
-                pageSize={pageSize}
-                totalItems={totalItems} />}
-            {!loading && !error && results.length > 0 && <Note message={noteText} />}
-            <Note message={(
-                    <>
-                        For &apos;All Awards&apos; we are showing the unique count of recipients across all award types, since some recipients receive multiple awards.
-                    </>
-            )} />
+            <SummaryInsightsContainer
+                // pass Recipient count to the summary section so we don't have to make the same API request again
+                resultsCount={tabCounts[activeTab]}
+                activeTab={activeTab}
+                areCountsLoading={inFlight}
+                overviewData={overviewData}
+                recipientOnly />
+            <RecipientTableContainer
+                activeTab={activeTab}
+                prevActiveTab={prevActiveTab}
+                scrollIntoView={scrollIntoViewTable} />
         </div>
     );
 };
 
-SpendingByRecipientContainer.propTypes = propTypes;
 export default SpendingByRecipientContainer;
